+++ src/services/content-org/content-org.js
@@ -1,0 +1,4 @@
+/**
+ * @namespace ContentOrganization
+ * @property {module:Playlists} Playlists
+ */