+++ CHANGELOG.md
@@ -2,6 +2,31 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [2.3.6](https://github.com/railroadmedia/musora-content-services/compare/v2.3.5...v2.3.6) (2025-04-04)
+
+### [2.3.5](https://github.com/railroadmedia/musora-content-services/compare/v2.3.4...v2.3.5) (2025-04-04)
+
+### [2.3.4](https://github.com/railroadmedia/musora-content-services/compare/v2.3.2...v2.3.4) (2025-04-03)
+
+### [2.3.3](https://github.com/railroadmedia/musora-content-services/compare/v2.3.2...v2.3.3) (2025-04-02)
+
+### [2.3.2](https://github.com/railroadmedia/musora-content-services/compare/v2.3.1...v2.3.2) (2025-03-31)
+
+### [2.3.1](https://github.com/railroadmedia/musora-content-services/compare/v2.3.0...v2.3.1) (2025-03-31)
+
+## [2.3.0](https://github.com/railroadmedia/musora-content-services/compare/v1.3.15...v2.3.0) (2025-03-31)
+
+
+### Features
+
+* **BEH-326:** update routing for top/all comment endpoints ([#203](https://github.com/railroadmedia/musora-content-services/issues/203)) ([1f4807f](https://github.com/railroadmedia/musora-content-services/commit/1f4807fcef3cb7ce0b03202519b4ccb4fbd6232f))
+* **MU2-357:** mock award data ([#204](https://github.com/railroadmedia/musora-content-services/issues/204)) ([008fea3](https://github.com/railroadmedia/musora-content-services/commit/008fea3e9aef75884fe0c3987bdbb42e9cb254da))
+
+
+### Bug Fixes
+
+* **playlists:** config import ([#198](https://github.com/railroadmedia/musora-content-services/issues/198)) ([4965de4](https://github.com/railroadmedia/musora-content-services/commit/4965de4fc2d1cda8de022ff46a3e09caac9d351f))
+
 ### [2.2.1](https://github.com/railroadmedia/musora-content-services/compare/v2.2.0...v2.2.1) (2025-03-28)
 
 ## [2.2.0](https://github.com/railroadmedia/musora-content-services/compare/v2.1.1...v2.2.0) (2025-03-27)